docker_images/micado-submitter/app/submitter/requirements.txt
ruamel.yaml==0.15.33
docker==2.5.1